+++ src/main/java/de/metas/ui/web/document/geo_location/GeoLocationAwareDescriptor.java
@@ -1,4 +1,4 @@
-package de.metas.ui.web.document.filter.provider.locationAreaSearch;
+package de.metas.ui.web.document.geo_location;
 
 import lombok.Builder;
 import lombok.NonNull;
@@ -28,7 +28,7 @@ import lombok.Value;
 
 @Value
 @Builder
-public final class LocationAreaSearchDescriptor
+public final class GeoLocationAwareDescriptor
 {
 	@NonNull
 	LocationColumnNameType type;
